refactor: move chat http module root
