+++ backend/chat/__init__.py
@@ -1,0 +1,1 @@
+"""Chat backend module root."""
