+++ backend/chat/api/http/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP routers for the chat backend."""
