+++ tests/Integration/test_auth_router.py
@@ -5,8 +5,8 @@ from types import SimpleNamespace
 import pytest
 from fastapi import HTTPException
 
+from backend.chat.api.http import router as chats_router
 from backend.web.routers import auth as auth_router
-from backend.web.routers import messaging as chats_router
 
 
 class _FakeAuthService:
