+++ tests/Integration/test_messaging_router.py
@@ -7,8 +7,8 @@ import pytest
 from fastapi import FastAPI, HTTPException
 from fastapi.testclient import TestClient
 
+from backend.chat.api.http import router as chat_router
 from backend.web.core.dependencies import get_app, get_current_user_id
-from backend.web.routers import messaging as messaging_router
 from backend.web.utils.serializers import avatar_url
 from storage.contracts import ContactEdgeRow
 
@@ -25,7 +25,7 @@ def _chat(chat_id: str) -> SimpleNamespace:
 def _route_test_app(state: SimpleNamespace) -> FastAPI:
     app = FastAPI()
     app.state = state
-    app.include_router(messaging_router.router)
+    app.include_router(chat_router.router)
     app.dependency_overrides[get_current_user_id] = lambda: "human-user-1"
     app.dependency_overrides[get_app] = lambda: app
     return app
@@ -89,20 +89,29 @@ def _create_chat_route_state(
 
 def test_messaging_crud_routes_are_sync_threadpool_boundaries() -> None:
     sync_routes = [
-        messaging_router.list_chats,
-        messaging_router.create_chat,
-        messaging_router.get_chat,
-        messaging_router.list_messages,
-        messaging_router.send_message,
-        messaging_router.retract_message,
-        messaging_router.delete_message_for_self,
-        messaging_router.mark_read,
-        messaging_router.delete_chat,
-        messaging_router.mute_chat,
+        chat_router.list_chats,
+        chat_router.create_chat,
+        chat_router.get_chat,
+        chat_router.list_messages,
+        chat_router.send_message,
+        chat_router.retract_message,
+        chat_router.delete_message_for_self,
+        chat_router.mark_read,
+        chat_router.delete_chat,
+        chat_router.mute_chat,
     ]
 
     assert [fn.__name__ for fn in sync_routes if inspect.iscoroutinefunction(fn)] == []
-    assert inspect.iscoroutinefunction(messaging_router.stream_chat_events)
+    assert inspect.iscoroutinefunction(chat_router.stream_chat_events)
+
+
+def test_chat_http_owner_module_lives_under_backend_chat() -> None:
+    from backend.web.routers import messaging as messaging_shell
+
+    assert chat_router.router is messaging_shell.router
+    assert chat_router.list_chats is messaging_shell.list_chats
+    assert chat_router.send_message is messaging_shell.send_message
+    assert chat_router.__name__ == "backend.chat.api.http.router"
 
 
 def test_get_accessible_chat_or_404_returns_chat():
@@ -114,7 +123,7 @@ def test_get_accessible_chat_or_404_returns_chat():
         )
     )
 
-    result = messaging_router._get_accessible_chat_or_404(app, "chat-1", "user-1")
+    result = chat_router._get_accessible_chat_or_404(app, "chat-1", "user-1")
 
     assert result is chat
 
@@ -128,7 +137,7 @@ def test_get_accessible_chat_or_404_raises_404_for_missing_chat():
     )
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router._get_accessible_chat_or_404(app, "missing", "user-1")
+        chat_router._get_accessible_chat_or_404(app, "missing", "user-1")
 
     assert exc_info.value.status_code == 404
     assert exc_info.value.detail == "Chat not found"
@@ -143,7 +152,7 @@ def test_get_accessible_chat_or_404_raises_403_for_non_member():
     )
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router._get_accessible_chat_or_404(app, "chat-1", "user-2")
+        chat_router._get_accessible_chat_or_404(app, "chat-1", "user-2")
 
     assert exc_info.value.status_code == 403
     assert exc_info.value.detail == "Not a participant of this chat"
@@ -163,7 +172,7 @@ def test_resolve_display_user_delegates_to_messaging_service(monkeypatch: pytest
         )
     )
 
-    result = messaging_router._resolve_display_user(app, "thread-user-1")
+    result = chat_router._resolve_display_user(app, "thread-user-1")
 
     assert result is expected
     assert seen == [("resolve_display_user", "thread-user-1")]
@@ -177,7 +186,7 @@ def test_get_chat_uses_access_helper(monkeypatch: pytest.MonkeyPatch):
         seen.append(("helper", (app_obj, chat_id, user_id)))
         return chat
 
-    monkeypatch.setattr(messaging_router, "_get_accessible_chat_or_404", fake_helper)
+    monkeypatch.setattr(chat_router, "_get_accessible_chat_or_404", fake_helper)
 
     app = SimpleNamespace(
         state=SimpleNamespace(
@@ -200,7 +209,7 @@ def test_get_chat_uses_access_helper(monkeypatch: pytest.MonkeyPatch):
         )
     )
 
-    result = messaging_router.get_chat("chat-1", user_id="user-1", app=app)
+    result = chat_router.get_chat("chat-1", user_id="user-1", app=app)
 
     assert result == {
         "id": "chat-1",
@@ -220,7 +229,7 @@ def test_delete_chat_uses_access_helper(monkeypatch: pytest.MonkeyPatch):
         seen.append(("helper", (app_obj, chat_id, user_id)))
         return chat
 
-    monkeypatch.setattr(messaging_router, "_get_accessible_chat_or_404", fake_helper)
+    monkeypatch.setattr(chat_router, "_get_accessible_chat_or_404", fake_helper)
 
     app = SimpleNamespace(
         state=SimpleNamespace(
@@ -231,7 +240,7 @@ def test_delete_chat_uses_access_helper(monkeypatch: pytest.MonkeyPatch):
         )
     )
 
-    result = messaging_router.delete_chat("chat-1", user_id="user-1", app=app)
+    result = chat_router.delete_chat("chat-1", user_id="user-1", app=app)
 
     assert result == {"status": "deleted"}
     assert seen == [
@@ -243,7 +252,7 @@ def test_delete_chat_uses_access_helper(monkeypatch: pytest.MonkeyPatch):
 def test_get_chat_resolves_thread_user_participant_via_thread_repo(monkeypatch: pytest.MonkeyPatch):
     chat = _chat("chat-1")
 
-    monkeypatch.setattr(messaging_router, "_get_accessible_chat_or_404", lambda _app, _chat_id, _user_id: chat)
+    monkeypatch.setattr(chat_router, "_get_accessible_chat_or_404", lambda _app, _chat_id, _user_id: chat)
 
     app = SimpleNamespace(
         state=SimpleNamespace(
@@ -278,7 +287,7 @@ def test_get_chat_resolves_thread_user_participant_via_thread_repo(monkeypatch: 
         )
     )
 
-    result = messaging_router.get_chat("chat-1", user_id="human-user-1", app=app)
+    result = chat_router.get_chat("chat-1", user_id="human-user-1", app=app)
 
     assert result["members"] == [
         {
@@ -332,7 +341,7 @@ def test_list_messages_resolves_thread_user_sender_name_via_thread_repo():
         )
     )
 
-    result = messaging_router.list_messages("chat-1", user_id="human-user-1", app=app)
+    result = chat_router.list_messages("chat-1", user_id="human-user-1", app=app)
 
     assert result == [
         {
@@ -442,9 +451,9 @@ def test_send_message_consumes_service_owned_message_projection() -> None:
         )
     )
 
-    result = messaging_router.send_message(
+    result = chat_router.send_message(
         "chat-1",
-        messaging_router.SendMessageBody(content="hello", sender_id="thread-user-1"),
+        chat_router.SendMessageBody(content="hello", sender_id="thread-user-1"),
         user_id="owner-user-1",
         app=app,
     )
@@ -519,9 +528,9 @@ def test_send_message_accepts_owned_thread_user_sender_id_via_thread_repo():
         )
     )
 
-    result = messaging_router.send_message(
+    result = chat_router.send_message(
         "chat-1",
-        messaging_router.SendMessageBody(content="hello", sender_id="thread-user-1"),
+        chat_router.SendMessageBody(content="hello", sender_id="thread-user-1"),
         user_id="owner-user-1",
         app=app,
     )
@@ -552,8 +561,8 @@ def test_create_chat_rejects_template_member_ids_for_group_participants() -> Non
     app = SimpleNamespace(state=state)
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router.create_chat(
-            messaging_router.CreateChatBody(
+        chat_router.create_chat(
+            chat_router.CreateChatBody(
                 user_ids=["human-user-1", "agent-user-1", "agent-user-2"],
                 title="bad-group",
             ),
@@ -575,8 +584,8 @@ def test_create_chat_rejects_template_member_id_for_direct_participant() -> None
     app = SimpleNamespace(state=state)
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router.create_chat(
-            messaging_router.CreateChatBody(
+        chat_router.create_chat(
+            chat_router.CreateChatBody(
                 user_ids=["human-user-1", "agent-user-1"],
                 title=None,
             ),
@@ -593,8 +602,8 @@ def test_create_chat_accepts_human_and_thread_social_user_ids_for_group_particip
     state, called = _create_chat_route_state(thread_user_ids={"thread-user-1", "thread-user-2"})
     app = SimpleNamespace(state=state)
 
-    result = messaging_router.create_chat(
-        messaging_router.CreateChatBody(
+    result = chat_router.create_chat(
+        chat_router.CreateChatBody(
             user_ids=["human-user-1", "thread-user-1", "thread-user-2"],
             title="good-group",
         ),
@@ -623,8 +632,8 @@ def test_create_group_chat_rejects_external_participant_without_active_relations
     app = SimpleNamespace(state=state)
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router.create_chat(
-            messaging_router.CreateChatBody(
+        chat_router.create_chat(
+            chat_router.CreateChatBody(
                 user_ids=["human-user-1", "owned-agent-1", "human-user-2"],
                 title="bad-group",
             ),
@@ -649,8 +658,8 @@ def test_create_group_chat_accepts_external_active_contact_without_relationship(
     )
     app = SimpleNamespace(state=state)
 
-    result = messaging_router.create_chat(
-        messaging_router.CreateChatBody(
+    result = chat_router.create_chat(
+        chat_router.CreateChatBody(
             user_ids=["human-user-1", "owned-agent-1", "human-user-2"],
             title="contact-group",
         ),
@@ -675,8 +684,8 @@ def test_create_group_chat_accepts_agent_owned_by_external_active_contact() -> N
     )
     app = SimpleNamespace(state=state)
 
-    result = messaging_router.create_chat(
-        messaging_router.CreateChatBody(
+    result = chat_router.create_chat(
+        chat_router.CreateChatBody(
             user_ids=["human-user-1", "owned-agent-1", "external-agent-1"],
             title="owner-contact-agent-group",
         ),
@@ -699,8 +708,8 @@ def test_create_group_chat_accepts_owned_agent_without_relationship() -> None:
     )
     app = SimpleNamespace(state=state)
 
-    result = messaging_router.create_chat(
-        messaging_router.CreateChatBody(
+    result = chat_router.create_chat(
+        chat_router.CreateChatBody(
             user_ids=["human-user-1", "owned-agent-1", "human-user-2"],
             title="good-group",
         ),
@@ -717,8 +726,8 @@ def test_create_chat_rejects_unknown_participant_ids_instead_of_falling_to_stora
     app = SimpleNamespace(state=state)
 
     with pytest.raises(HTTPException) as exc_info:
-        messaging_router.create_chat(
-            messaging_router.CreateChatBody(
+        chat_router.create_chat(
+            chat_router.CreateChatBody(
                 user_ids=["human-user-1", "thread-id-not-a-user", "thread-user-2"],
                 title="bad-group",
             ),
